Move code to sub components
