+++ client/src/app/pages/home/components/WatchedSbomDonutChart.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 
-
 import { SbomVulnerabilitiesDonutChart } from "@app/components/SbomVulnerabilitiesDonutChart";
 import { useSbomVulnerabilities } from "@app/hooks/domain-controls/useSbomVulnerabilities";
